// src/proto/dieter.ts
export interface User {
    /**
     * @generated from protobuf field: uint32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: string user_name = 2
     */
    userName: string;
    /**
     * @generated from protobuf field: string email = 3
     */
    email: string;
    /**
     * @generated from protobuf field: string picture = 4
     */
    picture: string;
}
/**
 * Post message
 *
 * @generated from protobuf message dieter.Post
 */
export interface Post {
    /**
     * @generated from protobuf field: uint32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: uint32 user_id = 2
     */
    userId: number;
    /**
     * @generated from protobuf field: string content = 3
     */
    content: string;
    /**
     * @generated from protobuf field: string image_url = 4
     */
    imageUrl: string;
    /**
     * @generated from protobuf field: repeated string images = 5
     */
    images: string[];
    /**
     * @generated from protobuf field: bool is_public = 6
     */
    isPublic: boolean;
    /**
     * @generated from protobuf field: bool is_sensitive = 7
     */
    isSensitive: boolean;
    /**
     * @generated from protobuf field: string author_name = 8
     */
    authorName: string;
    /**
     * @generated from protobuf field: string author_picture = 9
     */
    authorPicture: string;
    /**
     * @generated from protobuf field: bool is_retweet = 10
     */
    isRetweet: boolean;
    /**
     * @generated from protobuf field: uint32 retweet_user_id = 11
     */
    retweetUserId: number;
    /**
     * @generated from protobuf field: string retweet_user_name = 12
     */
    retweetUserName: string;
    /**
     * @generated from protobuf field: string retweet_user_picture = 13
     */
    retweetUserPicture: string;
    /**
     * @generated from protobuf field: string retweeted_at = 14
     */
    retweetedAt: string;
    /**
     * @generated from protobuf field: dieter.User user = 15
     */
    user?: User;
    /**
     * @generated from protobuf field: repeated dieter.Comment comments = 16
     */
    comments: Comment[];
    /**
     * @generated from protobuf field: repeated dieter.Retweet retweets = 17
     */
    retweets: Retweet[];
    /**
     * @generated from protobuf field: repeated dieter.Like likes = 18
     */
    likes: Like[];
    /**
     * @generated from protobuf field: string created_at = 19
     */
    createdAt: string;
    /**
     * @generated from protobuf field: string updated_at = 20
     */
    updatedAt: string;
}
/**
 * Comment message
 *
 * @generated from protobuf message dieter.Comment
 */
export interface Comment {
    /**
     * @generated from protobuf field: uint32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: uint32 post_id = 2
     */
    postId: number;
    /**
     * @generated from protobuf field: uint32 user_id = 3
     */
    userId: number;
    /**
     * @generated from protobuf field: string content = 4
     */
    content: string;
    /**
     * @generated from protobuf field: string author_name = 5
     */
    authorName: string;
    /**
     * @generated from protobuf field: string author_picture = 6
     */
    authorPicture: string;
    /**
     * @generated from protobuf field: dieter.User user = 7
     */
    user?: User;
    /**
     * @generated from protobuf field: string created_at = 8
     */
    createdAt: string;
    /**
     * @generated from protobuf field: string updated_at = 9
     */
    updatedAt: string;
}
/**
 * Retweet message
 *
 * @generated from protobuf message dieter.Retweet
 */
export interface Retweet {
    /**
     * @generated from protobuf field: uint32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: uint32 post_id = 2
     */
    postId: number;
    /**
     * @generated from protobuf field: uint32 user_id = 3
     */
    userId: number;
    /**
     * @generated from protobuf field: dieter.User user = 4
     */
    user?: User;
    /**
     * @generated from protobuf field: string created_at = 5
     */
    createdAt: string;
    /**
     * @generated from protobuf field: string updated_at = 6
     */
    updatedAt: string;
}
/**
 * Like message
 *
 * @generated from protobuf message dieter.Like
 */
export interface Like {
    /**
     * @generated from protobuf field: uint32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: uint32 post_id = 2
     */
    postId: number;
    /**
     * @generated from protobuf field: uint32 user_id = 3
     */
    userId: number;
    /**
     * @generated from protobuf field: dieter.User user = 4
     */
    user?: User;
    /**
     * @generated from protobuf field: string created_at = 5
     */
    createdAt: string;
    /**
     * @generated from protobuf field: string updated_at = 6
     */
    updatedAt: string;
}
// ...
export interface Pagination {
    /**
     * @generated from protobuf field: uint32 page = 1
     */
    page: number;
    /**
     * @generated from protobuf field: uint32 limit = 2
     */
    limit: number;
    /**
     * @generated from protobuf field: uint32 total = 3
     */
    total: number;
}
// ...
export interface PostsResponse {
    /**
     * @generated from protobuf field: repeated dieter.Post posts = 1
     */
    posts: Post[];
    /**
     * @generated from protobuf field: dieter.Pagination pagination = 2
     */
    pagination?: Pagination;
}
// ...
export const User = new User$Type();
// ...
export const Post = new Post$Type();
// ...
export const Comment = new Comment$Type();
// ...
export const Retweet = new Retweet$Type();
// ...
export const Like = new Like$Type();
// ...
class Pagination$Type extends MessageType<Pagination> {
    constructor() {
        super("dieter.Pagination", [
            { no: 1, name: "page", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 2, name: "limit", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 3, name: "total", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<Pagination>): Pagination {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.page = 0;
        message.limit = 0;
        message.total = 0;
        if (value !== undefined)
            reflectionMergePartial<Pagination>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Pagination): Pagination {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* uint32 page */ 1:
                    message.page = reader.uint32();
                    break;
                case /* uint32 limit */ 2:
                    message.limit = reader.uint32();
                    break;
                case /* uint32 total */ 3:
                    message.total = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Pagination, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* uint32 page = 1; */
        if (message.page !== 0)
            writer.tag(1, WireType.Varint).uint32(message.page);
        /* uint32 limit = 2; */
        if (message.limit !== 0)
            writer.tag(2, WireType.Varint).uint32(message.limit);
        /* uint32 total = 3; */
        if (message.total !== 0)
            writer.tag(3, WireType.Varint).uint32(message.total);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message dieter.Pagination
 */
export const Pagination = new Pagination$Type();
// ...
class PostsResponse$Type extends MessageType<PostsResponse> {
    constructor() {
        super("dieter.PostsResponse", [
            { no: 1, name: "posts", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => Post },
            { no: 2, name: "pagination", kind: "message", T: () => Pagination }
        ]);
    }
    create(value?: PartialMessage<PostsResponse>): PostsResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.posts = [];
        if (value !== undefined)
            reflectionMergePartial<PostsResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PostsResponse): PostsResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated dieter.Post posts */ 1:
                    message.posts.push(Post.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* dieter.Pagination pagination */ 2:
                    message.pagination = Pagination.internalBinaryRead(reader, reader.uint32(), options, message.pagination);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PostsResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated dieter.Post posts = 1; */
        for (let i = 0; i < message.posts.length; i++)
            Post.internalBinaryWrite(message.posts[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* dieter.Pagination pagination = 2; */
        if (message.pagination)
            Pagination.internalBinaryWrite(message.pagination, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message dieter.PostsResponse
 */
export const PostsResponse = new PostsResponse$Type();